+++ sdk/core/azure_core/src/error/mod.rs
@@ -1,6 +1,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-//! Client-specific error functions.
+pub use typespec_client_core::error::*;
+mod error_response;
 
-pub use typespec::error::*;
+pub use error_response::*;