+++ sdk/core/azure_core/src/lib.rs
@@ -11,9 +11,9 @@ mod macros;
 
 mod constants;
 pub mod credentials;
+pub mod error;
 pub mod hmac;
 pub mod http;
-
 #[cfg(feature = "test")]
 pub mod test;
 
@@ -21,9 +21,8 @@ pub use constants::*;
 
 // Re-export modules in typespec_client_core such that azure_core-based crates don't need to reference it directly.
 pub use typespec_client_core::{
-    async_runtime, base64, create_enum, create_extensible_enum,
-    error::{self, Error, Result},
-    fmt, json, sleep, stream, time, Bytes, Uuid,
+    async_runtime, base64, create_enum, create_extensible_enum, fmt, json, sleep, stream, time,
+    Bytes, Error, Result, Uuid,
 };
 
 /// Abstractions for distributed tracing and telemetry.