+++ sdk/typespec/typespec_client_core/src/http/options/mod.rs
@@ -72,7 +72,6 @@ impl Default for PipelineOptions {
         Self {
             retry_headers: RetryHeaders {
                 retry_headers: vec![RETRY_AFTER],
-                error_header: None,
             },
         }
     }